requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.4
Babel==2.6.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bootstrap4==0.1.0
CacheControl==0.12.5
cachetools==3.1.1
celery==4.4.3
celery-stubs==0.1.2
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
credentials==1.1
dataclasses==0.6
Django==2.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-constance==2.4.0
django-crontab==0.7.1
django-db==0.0.7
django-debug-toolbar==1.11
django-filter==2.1.0
django-jalali==3.1.0
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.9.1
django-multiselectfield==0.1.8
django-sendgrid-webhook==0.3.0
django-stubs==0.9.0
django-timezone-field==3.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
dnspython==1.16.0
eventlet==0.30.2
firebase-admin==3.0.0
Flask==1.1.2
flower==0.9.3
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-crc32c==1.1.2
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==1.0.0
grpcio==1.24.1
httplib2==0.14.0
humanize==3.4.1
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
jdatetime==3.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
messaging==1.1
more-itertools==7.2.0
msgpack==0.6.2
mypy==0.720
mypy-extensions==0.4.1
mysqlclient==1.4.4
packaging==20.9
Pillow==5.4.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.12.0
psycopg2==2.7.7
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
PySocks==1.6.8
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==2.2.1
sqlparse==0.3.0
tornado==5.1.1
twilio==6.25.2
typed-ast==1.4.3
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==0.6.0
